requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
pandas>=1.2.0
fake-useragent>=0.1.11
extruct>=0.13.0
w3lib>=1.22.0
price-parser>=0.3.4
langchain>=0.0.200
langchain-community>=0.0.10
openai>=0.27.0
python-dotenv>=0.19.0
pydantic>=1.8.2 